requirements.txt
torch~=2.1.0.dev20230819+cu121
torchvision~=0.16.0.dev20230819+cu121
transformers==4.5.1
addict~=2.4.0
yapf~=0.40.1
timm~=0.9.5
numpy~=1.24.1
opencv-python~=4.8.0.76
supervision==0.3.2
pycocotools~=2.0.7
packaging==21.3
ftfy
regex
tqdm~=4.66.1
PyYAML~=6.0.1
termcolor~=2.3.0
Pillow~=9.5.0
matplotlib~=3.7.2
setuptools~=65.5.1
requests~=2.31.0
clip~=1.0